Rename integration tests to add IT suffix
