+++ build.gradle
@@ -14,6 +14,7 @@ buildscript {
 plugins {
   id 'application'
   id 'checkstyle'
+  id 'idea'
   id 'pmd'
   id 'jacoco'
   id 'org.owasp.dependencycheck' version '6.2.2'
@@ -49,6 +50,13 @@ sourceSets {
       runtimeClasspath += main.output
       srcDir file('src/integrationTest/java')
     }
+  }
+}
+
+idea {
+  module {
+    testSourceDirs += file('src/integrationTest/java')
+    testSourceDirs += file('src/functionalTest/java')
   }
 }
