+++ src/test/java/uk/gov/hmcts/divorce/testutil/TestDataHelper.java
@@ -7,6 +7,7 @@ import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import feign.FeignException;
 import feign.Request;
 import feign.Response;
+import uk.gov.hmcts.ccd.sdk.type.AddressGlobalUK;
 import uk.gov.hmcts.ccd.sdk.type.Document;
 import uk.gov.hmcts.ccd.sdk.type.DynamicList;
 import uk.gov.hmcts.ccd.sdk.type.DynamicListElement;
@@ -111,6 +112,21 @@ public class TestDataHelper {
             .gender(gender)
             .languagePreferenceWelsh(NO)
             .contactDetailsConfidential(SHARE)
+            .build();
+    }
+
+    public static Applicant getApplicantWithAddress() {
+        return Applicant.builder()
+            .firstName(TEST_FIRST_NAME)
+            .middleName(TEST_MIDDLE_NAME)
+            .lastName(TEST_LAST_NAME)
+            .email(TEST_USER_EMAIL)
+            .languagePreferenceWelsh(NO)
+            .homeAddress(AddressGlobalUK.builder()
+                .addressLine1("line 1")
+                .postTown("town")
+                .postCode("postcode")
+                .build())
             .build();
     }
